+++ dev/cache.js
@@ -1,93 +1,66 @@
 "use strict";
 
 let chalk = require("chalk");
-
-let utils = require("../../src/utils");
-let ServiceBroker = require("../../src/service-broker");
+let ServiceBroker = require("../src/service-broker");
 
 // Create broker
 let broker = new ServiceBroker({
-	namespace: "demo",
-	transporter: null,
-	logger: console,
 	logLevel: "debug",
-	metrics: true,
-	cacher: true,
+	cacher: {
+		type: "MemoryLRU",
+		options: {
+			max: 100,
+			ttl: 3
+		}
+	}
 });
 
-broker.loadService("./examples/post.service.js");
-broker.loadService("./examples/user.service.js");
-
-
-console.log(chalk.bold(">> Get all users"));
+broker.createService({
+	name: "greeter",
+	actions: {
+		hello: {
+			//cache: false,
+			cache: {
+				//enabled: true
+				enabled: ctx => ctx.params.noCache !== true
+			},
+			handler(ctx) {
+				this.logger.debug(chalk.yellow("Execute handler"));
+				return `Hello ${ctx.params.name}`;
+			}
+		}
+	}
+});
 
 broker.start()
-	.then(() => {
-		return broker.call("v2.users.find");
-	})
-
-	.then(data => {
-		console.log("v2.users.find response length:", data.length, "\n");
-	})
+	.then(() => broker.call("greeter.hello", { name: "Moleculer" }))
+	.then(() => broker.call("greeter.hello", { name: "Moleculer" }))
+	.then(() => broker.call("greeter.hello", { name: "Moleculer", noCache: true }))
+	.then(() => broker.call("greeter.hello", { name: "Moleculer" }))
+	.then(() => broker.call("greeter.hello", { name: "Moleculer" }, { meta: { $cache: false }}))
 
 	.then(() => {
-		console.log(chalk.bold(">> Get user.5 (not found in the cache)"));
-		return broker.call("v2.users.get", {
-			id: 5
-		});
-	})
-	.then(data => {
-		console.log("v2.users.get(5) response user email:", data.email, "\n");
+		for(let i = 0; i < 1000; i++) {
+			broker.cacher.set(`key-${i}`, i);
+			if (i % 10 == 0) {
+				broker.cacher.get(`key-${100}`);
+				broker.cacher.get(`key-${200}`);
+				broker.cacher.get(`key-${500}`);
+				broker.cacher.get(`key-${400}`);
+				broker.cacher.get(`key-${300}`);
+			}
+		}
+
+		broker.logger.info("Length:", broker.cacher.cache.length);
+		broker.logger.info("keys:", broker.cacher.cache.keys().join(","));
 	})
 
+	.delay(5 * 1000)
 	.then(() => {
-		console.log(chalk.bold(">> Get all posts (populate authors from users"));
-		return broker.call("posts.find", {
-			limit: 10
-		});
-	})
-	.then(data => {
-		console.log("posts.find response length:", data.length, "\n");
-
-		console.log(chalk.bold(">> Get posts.4 (populate author from cache)"));
-
-		return broker.call("posts.get", {
-			id: data[4].id
-		}).then((post) => {
-			console.log("posts[4].author email:", post.author.email, "\n");
-		});
+		broker.logger.info("=========================================");
+		broker.logger.info("Length:", broker.cacher.cache.length);
+		broker.logger.info("keys:", broker.cacher.cache.keys().join(","));
 	})
 
-	.then(() => {
-		// Get from cache
-		console.log(chalk.bold(">> Get user.5 again (found in the cache)"));
-		return broker.call("v2.users.get", {
-			id: 5
-		});
-	})
-	.then(data => {
-		console.log("v2.users.get(5) (cache) user email:", data.email, "\n");
-	})
-
-	.then(() => {
-		console.log(chalk.yellow.bold("CLEAN CACHE: v2.users.*\n"));
-		// Clear the cache
-		return broker.cacher.clean("v2.users.*");
-	})
-	.then(utils.delay(100))
-	.then(() => {
-		console.log(chalk.bold("\n>> Get user.5 again (not found in the cache, after clean)"));
-		// Not found in the cache
-		return broker.call("v2.users.get", {
-			id: 5
-		});
-	})
-	.then(data => {
-		console.log("v2.users.get(5) response user email:", data.email, "\n");
-	})
-
-	.catch((err) => {
-		console.error("Error!", err);
-	})
-
+	.catch(err => broker.logger.error(err))
 	.then(() => broker.stop());