+++ src/cachers/base.js
@@ -92,7 +92,7 @@ class Cacher {
 	/**
 	 * Delete a content by key from cache
 	 *
-	 * @param {any} key
+	 * @param {string|Array<string>} key
 	 *
 	 * @memberof Cacher
 	 */
@@ -104,7 +104,7 @@ class Cacher {
 
 	/**
 	 * Clean cache. Remove every key by match
-	 * @param {any} match string. Default is "**"
+	 * @param {string|Array<string>} match string. Default is "**"
 	 * @returns {Promise}
 	 *
 	 * @memberof Cacher
@@ -217,9 +217,23 @@ class Cacher {
 	 */
 	middleware() {
 		return (handler, action) => {
-			if (action.cache) {
+			const opts = _.defaultsDeep({}, _.isObject(action.cache) ? action.cache : { enabled: !!action.cache });
+			if (opts.enabled !== false) {
+				const isEnabledFunction = _.isFunction(opts.enabled);
+
 				return function cacherMiddleware(ctx) {
-					const cacheKey = this.getCacheKey(action.name, ctx.params, ctx.meta, action.cache.keys);
+					if (isEnabledFunction) {
+						if (!opts.enabled.call(ctx.service, ctx)) {
+							// Cache is disabled. Call the handler only.
+							return handler(ctx);
+						}
+					}
+
+					// Disable caching with `ctx.meta.$cache = false`
+					if (ctx.meta["$cache"] === false)
+						return handler(ctx);
+
+					const cacheKey = this.getCacheKey(action.name, ctx.params, ctx.meta, opts.keys);
 					return this.get(cacheKey).then(content => {
 						if (content != null) {
 							// Found in the cache! Don't call handler, return with the content
@@ -230,7 +244,7 @@ class Cacher {
 						// Call the handler
 						return handler(ctx).then(result => {
 							// Save the result to the cache
-							this.set(cacheKey, result, action.cache.ttl);
+							this.set(cacheKey, result, opts.ttl);
 
 							return result;
 						});