+++ src/service-broker.js
@@ -335,7 +335,7 @@ class ServiceBroker {
 				this.logger.info(`ServiceBroker with ${this.services.length} service(s) is started successfully.`);
 				this.started = true;
 
-				this.localBus.emit("$broker.started");
+				this.broadcastLocal("$broker.started");
 			})
 			.then(() => {
 				if (this.transit)
@@ -400,7 +400,7 @@ class ServiceBroker {
 			.then(() => {
 				this.logger.info("ServiceBroker is stopped. Good bye.");
 
-				this.localBus.emit("$broker.stopped");
+				this.broadcastLocal("$broker.stopped");
 
 				if ((this.options.skipProcessEventRegistration || false) !== true) {
 					process.removeListener("beforeExit", this._closeFn);
@@ -509,11 +509,9 @@ class ServiceBroker {
 		else
 			serviceFiles = glob.sync(path.join(folder, fileMask));
 
-		if (serviceFiles) {
-			serviceFiles.forEach(filename => {
-				this.loadService(filename);
-			});
-		}
+		if (serviceFiles)
+			serviceFiles.forEach(filename => this.loadService(filename));
+
 		return serviceFiles.length;
 	}
 
@@ -535,6 +533,7 @@ class ServiceBroker {
 			schema = require(fName);
 		} catch (e) {
 			this.logger.error(`Failed to load service '${fName}'`, e);
+			throw e;
 		}
 
 		let svc;
@@ -674,7 +673,6 @@ class ServiceBroker {
 	 */
 	registerLocalService(registryItem) {
 		this.registry.registerLocalService(registryItem);
-		this.servicesChanged(true);
 	}
 
 	/**
@@ -932,8 +930,10 @@ class ServiceBroker {
 		if (opts.ctx != null) {
 			// Reused context
 			ctx = opts.ctx;
-			ctx.endpoint = endpoint;
-			ctx.action = endpoint.action;
+			if (endpoint) {
+				ctx.endpoint = endpoint;
+				ctx.action = endpoint.action;
+			}
 		} else {
 			// New root context
 			ctx = this.ContextFactory.create(this, endpoint, params, opts);
@@ -1101,14 +1101,27 @@ class ServiceBroker {
 		this.logger.debug(`Broadcast '${eventName}' event`+ (groups ? ` to '${groups.join(", ")}' group(s)` : "") + ".");
 
 		if (this.transit) {
-			const endpoints = this.registry.events.getAllEndpoints(eventName, groups);
+			if (!this.options.disableBalancer) {
+				const endpoints = this.registry.events.getAllEndpoints(eventName, groups);
 
-			// Send to remote services
-			endpoints.forEach(ep => {
-				if (ep.id != this.nodeID) {
-					return this.transit.sendBroadcastEvent(ep.id, eventName, payload, groups);
+				// Send to remote services
+				endpoints.forEach(ep => {
+					if (ep.id != this.nodeID) {
+						return this.transit.sendBroadcastEvent(ep.id, eventName, payload, groups);
+					}
+				});
+			} else {
+				// Disabled balancer case
+				if (!groups || groups.length == 0) {
+					// Apply to all groups
+					groups = this.getEventGroups(eventName);
 				}
-			});
+
+				if (groups.length == 0)
+					return;
+
+				return this.transit.sendBroadcastEvent(null, eventName, payload, groups);
+			}
 		}
 
 		// Send to local services