+++ test/integration/registry.spec.js
@@ -85,7 +85,13 @@ const mailService = {
 };
 
 describe("Test service registry", () => {
-	const master = H.createNode("first", "master", []);
+	const servicesChanged = jest.fn();
+	const master = H.createNode("first", "master", [{
+		name: "watcher",
+		events: {
+			"$services.changed": servicesChanged
+		}
+	}]);
 	const node1 = H.createNode("first", "node-1", [userService]);
 
 	let node2 = H.createNode("first", "node-2", [userService, paymentService]);
@@ -97,9 +103,12 @@ describe("Test service registry", () => {
 		expect(H.hasService(master, "$node")).toBe(true);
 		expect(H.hasService(master, "users")).toBe(false);
 		expect(H.getActionNodes(master, "$node.list")).toEqual(["master"]);
+		expect(servicesChanged).toHaveBeenCalledTimes(1);
 	});
 
 	it("start node1 with userService", () => {
+		servicesChanged.mockClear();
+
 		return node1.start().delay(100).then(() => {
 			expect(H.getNode(master, "node-1")).toBeDefined();
 			expect(H.hasService(master, "users")).toBe(true);
@@ -118,11 +127,15 @@ describe("Test service registry", () => {
 			expect(H.getActionNodes(master, "$node.list")).toEqual(["master", "node-1"]);
 
 			expect(H.getEventNodes(master, "user.created")).toEqual(["node-1"]);
+
+			expect(servicesChanged).toHaveBeenCalledTimes(1);
 		});
 	});
 
 
 	it("start node2 with userService & payment service", () => {
+		servicesChanged.mockClear();
+
 		return node2.start().delay(100).then(() => {
 			let node2 = H.getNode(master, "node-2");
 			expect(node2).toBeDefined();
@@ -141,10 +154,14 @@ describe("Test service registry", () => {
 
 			expect(H.getEventNodes(master, "user.created")).toEqual(["node-1", "node-2"]);
 			expect(H.getEventNodes(master, "user.paid")).toEqual(["node-2"]);
+
+			expect(servicesChanged).toHaveBeenCalledTimes(1);
 		});
 	});
 
 	it("stop node2", () => {
+		servicesChanged.mockClear();
+
 		return node2.stop().delay(100).then(() => {
 			let infoNode2 = H.getNode(master, "node-2");
 			expect(infoNode2).toBeDefined();
@@ -161,10 +178,14 @@ describe("Test service registry", () => {
 
 			expect(H.getEventNodes(master, "user.created")).toEqual(["node-1"]);
 			expect(H.getEventNodes(master, "user.paid")).toEqual([]);
+
+			expect(servicesChanged).toHaveBeenCalledTimes(1);
 		});
 	});
 
 	it("node2 recreate with posts, paymentMod", () => {
+		servicesChanged.mockClear();
+
 		node2 = H.createNode("first", "node-2", [paymentModService, postService]);
 
 		return node2.start().delay(100).then(() => {
@@ -184,24 +205,30 @@ describe("Test service registry", () => {
 			expect(H.getEventNodes(master, "user.created")).toEqual(["node-1", "node-2"]);
 			expect(H.getEventNodes(master, "user.paid")).toEqual([]);
 			expect(H.getEventNodes(master, "payment.done")).toEqual(["node-2"]);
+
+			expect(servicesChanged).toHaveBeenCalledTimes(1);
 		});
 	});
 
 	it("load mail on-the-fly to node1", () => {
+		servicesChanged.mockClear();
 		H.addServices(node1, [mailService]);
 
 		return Promise.resolve().delay(100).then(() => {
 			expect(H.hasService(master, "mail")).toBe(true);
 			expect(H.getActionNodes(master, "mail.send")).toEqual(["node-1"]);
+			expect(servicesChanged).toHaveBeenCalledTimes(1);
 		});
 	});
 
 	it("destroy mail on-the-fly to node1", () => {
+		servicesChanged.mockClear();
 		H.removeServices(node1, ["mail"]);
 
 		return Promise.resolve().delay(100).then(() => {
 			expect(H.hasService(master, "mail")).toBe(false);
 			expect(H.getActionNodes(master, "mail.send")).toEqual([]);
+			expect(servicesChanged).toHaveBeenCalledTimes(1);
 		});
 	});
 });